Cleanup & added endpoint for user fetch
